backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
numpy==1.26.4
pillow==10.3.0
scikit-learn==1.4.2
joblib==1.4.0
